Buggy bitmap index join
When one of the joined table has bitmap index, we
would use bitmap index join

Algorithm:
Suppose we are joining table A and B, where we
want to use bitmap index on B:
1. probe: probe each row in A by key 'k'
2. match: fetch matches of 'k' in B using index
3. combine: join rows from two tables

WARNING:
This is a BUGGY implementation.
1. Failed on TPCH 7, 9, 10, 11, 20, 21
2. No runtime error for now.

A simple testcase 'hash_join_basic.sql' is
attached for sanity check and performance check.